cli/pyproject.toml
[project]
name = "gitpoli-cli"
version = "0.1.0"
description = "CLI for gitpoli framework"
authors = ["Tu Nombre <tu.email@ejemplo.com>"]
readme = "README.md"
requires-python = ">=3.8"

[tool.poetry.dependencies]
python = ">=3.8"
typer = "^0.9.0"
rich = "^13.7.0"
pyyaml = "^6.0.1"

[project.scripts]
gitpoli = "gitpoli_cli.main:app"
